Delay pools 64-bit buckets and IPv6-polish
This patch adds some polish and minor unfinished business in IPv6 
support vs delay pools.

Also updates the bucket size to 64-bytes to allow for >4MB traffic caps. 

TODO: Bump up the refresh size for >4MB/sec flows and check that nothing
      gets broken by it.

+++ src/DelaySpec.h
@@ -32,6 +32,9 @@
 #ifndef SQUID_DELAYSPEC_H
 #define SQUID_DELAYSPEC_H
 
+/* for int64_t definition */
+#include "config.h"
+
 class StoreEntry;
 
 /// \ingroup DelyPoolsAPI
@@ -44,7 +47,7 @@ public:
     void dump(StoreEntry *) const;
     void parse();
     int restore_bps;
-    int max_bytes;
+    int64_t max_bytes;
 };
 
 #endif /* SQUID_DELAYSPEC_H */
